@@ -1,35 +1,32 @@
-# -*- coding: utf-8 -*-
 """
 Index
 =====
 """
 
+from dash.dcc import Link, Location, Markdown
 from dash.dependencies import Input, Output
+from dash.html import A, Div, H3, Main, P
 from dash_bootstrap_components import Col, Container, Row
-from dash_core_components import Link, Location, Markdown
-from dash_html_components import A, Div, H3, Main, P
 
 import apps.idt_calculator as app_1
 from app import APP, SERVER  # noqa
 
-__author__ = 'Alex Forsythe, Gayle McAdams, Thomas Mansencal, Nick Shaw'
-__copyright__ = ('Copyright (C) 2020-2021 '
-                 'Academy of Motion Picture Arts and Sciences')
-__license__ = 'Academy of Motion Picture Arts and Sciences License Terms'
-__maintainer__ = 'Academy of Motion Picture Arts and Sciences'
-__email__ = 'acessupport@oscars.org'
-__status__ = 'Production'
+__author__ = "Alex Forsythe, Gayle McAdams, Thomas Mansencal, Nick Shaw"
+__copyright__ = "Copyright 2020 Academy of Motion Picture Arts and Sciences"
+__license__ = "Academy of Motion Picture Arts and Sciences License Terms"
+__maintainer__ = "Academy of Motion Picture Arts and Sciences"
+__email__ = "acessupport@oscars.org"
+__status__ = "Production"
 
-__all__ = ['load_app']
+__all__ = ["load_app"]
 
-APP.layout = Container([Location(id='url', refresh=False), Div(id='apps')])
+APP.layout = Container([Location(id="url", refresh=False), Div(id="apps")])
 
 
-@APP.callback(Output('apps', 'children'), [Input('url', 'pathname')])
+@APP.callback(Output("apps", "children"), [Input("url", "pathname")])
 def load_app(app):
     """
-    Callback loading given app into the appropriate :class:`Div` class
-    instance.
+    Load given app into the appropriate :class:`Div` class instance.
 
     Parameters
     ----------
@@ -45,29 +42,48 @@ def load_app(app):
     if app == app_1.APP_PATH:
         return app_1.LAYOUT
     else:
-        return Container([
-            Main([
-                Row([
-                    Col([
-                        P([
-                            'Various A.M.P.A.S. colour science ',
-                            A('Dash',
-                              href='https://dash.plot.ly/',
-                              target='_blank'), ' apps.'
-                        ]),
-                        H3([
-                            Link(
-                                app_1.APP_NAME,
-                                href=app_1.APP_PATH,
-                                className='app-link')
-                        ]),
-                        Markdown(
-                            app_1.APP_DESCRIPTION.replace('This app c', 'C')),
-                    ])
-                ])
-            ])
-        ])
+        return Container(
+            [
+                Main(
+                    [
+                        Row(
+                            [
+                                Col(
+                                    [
+                                        P(
+                                            [
+                                                "Various A.M.P.A.S. colour science ",
+                                                A(
+                                                    "Dash",
+                                                    href="https://dash.plot.ly/",
+                                                    target="_blank",
+                                                ),
+                                                " apps.",
+                                            ]
+                                        ),
+                                        H3(
+                                            [
+                                                Link(
+                                                    app_1.APP_NAME,
+                                                    href=app_1.APP_PATH,
+                                                    className="app-link",
+                                                )
+                                            ]
+                                        ),
+                                        Markdown(
+                                            app_1.APP_DESCRIPTION.replace(
+                                                "This app c", "C"
+                                            )
+                                        ),
+                                    ]
+                                )
+                            ]
+                        )
+                    ]
+                )
+            ]
+        )
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     APP.run_server(debug=True)
